Added new tools for preprocessing and enabled multiview fusion
Extended preprocessing capabilities:

* Processing og ome.tiff files:

PROBLEM:OpenSPIM uses .ome.tiff format.

SOLUTION: Conversion of .ome.tif into .tif is now supported.

* Processing of discontinous .czi datasets:

PROBLEM: When restarting the imaging in ZEN for a recording the
resulting .czi files are discontinous and treated as different
dataset. Thus one cannot define 1 dataset in the first processing
step of Multiview Reconstruction.

SOLUTION: To address that the .czi files are first rename into a
conected pattern and then resaved into .tif or .zip

LIMITATION: The resaving of .czi files is still based on lifeline
version of Fiji containing an old bioformats version.

* Compression:

PROBLEM: limitation of storage space for transient files of
processing

SOLUTION: Resave into a compressed file using resave to ZIP
in FIJI. This replaces the resave to .tif scripts. Allows a almost
2x reduction in the size of the files and is promised to speed up
the downstream processing

Added fusion to multiview processing:

* Added multiview fusion to pipeline:

LIMITATION: The fusion is not cluster enabled and writes the output
files as .tif. It does not write a new .xml file or add to the
existing file. Thus there is no support via the BigDataViewer yet.