Borught hdf5 resave to new plugin version.
BUG FIX:  When dataset is defined from .czi the timepoints start
from 0. The First job however in the resave as hdf5 is also 0.
Therefore for hdf5 resaving parallel_timepoints need to
be ="seq 1 2" not ="seq 0 2" but for the rest of the pipeline
it is ="seq 0 2". NOW: create-export-jobs.sh calculates the
number of needed partitions based on parallel_timepoints. Then
automatically sets the first timepoint to 1. Jobs 0 still creates
the .xml file.